Adding an HTML reporter for in browser reporting
This reporter heavily uses Backbone and underscore/lodash for rendering
an HTML report in a <div id="coverage"></coverage>. It also renders a
menu of files in a <div id="menu"></div>. Coverage is rendered in a
table, not within pre tags as nodejs reporters from CoverJS.

This Reporter can be adapted to be used with any testing framework such
as Jasmine, QUnit or Mocha. This one here, produces HTML to be used
with some Mocha CSS and js files, for better usability.